+++ go.mod
@@ -5,5 +5,5 @@ go 1.16
 require (
 	github.com/jlegrone/xbargo v0.0.0-20220128073828-b95b21d50723
 	github.com/urfave/cli/v2 v2.3.0
-	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a
 )
